+++ src/embeddingLib/include/embeddingIO/EmbeddingIO.hpp
@@ -24,7 +24,7 @@ enum EmbeddingType {
 class EmbeddingIO {
    public:
     static std::unique_ptr<Embedding> parseEmbedding(EmbeddingType type,
-                                                     const std::vector<std::vector<double>>& coordinates, int lpNorm);
+                                                     const std::vector<std::vector<float>>& coordinates, int lpNorm);
 
     /**
      * Reads coordinates for a graph from a file.
@@ -35,7 +35,7 @@ class EmbeddingIO {
      *
      * Assumes the ids in the file to be consecutive starting from 0.
      */
-    static std::vector<std::vector<double>> readCoordinatesFromFile(std::string filePath, std::string comment = "#",
+    static std::vector<std::vector<float>> readCoordinatesFromFile(std::string filePath, std::string comment = "#",
                                                                     std::string delimiter = ",");
 
     /**
@@ -43,19 +43,19 @@ class EmbeddingIO {
      *
      * Splits the last column away and writes it into a new vector
      */
-    static std::pair<std::vector<std::vector<double>>, std::vector<double>> splitLastColumn(
-        const std::vector<std::vector<double>>& coordinates);
+    static std::pair<std::vector<std::vector<float>>, std::vector<float>> splitLastColumn(
+        const std::vector<std::vector<float>>& coordinates);
 
     /**
      * Used to get the weights for mercator embeddings.
      */
-    static std::pair<std::vector<double>, std::vector<std::vector<double>>> splitFirstColumn(
-        const std::vector<std::vector<double>>& coordinates);
+    static std::pair<std::vector<float>, std::vector<std::vector<float>>> splitFirstColumn(
+        const std::vector<std::vector<float>>& coordinates);
 
     /**
      * mapping maps NodeIds of the input file to nodeIds in the position vector
      */
-    static void writeCoordinates(std::string filePath, const std::vector<std::vector<double>>& positions,
-                                 const std::vector<double>& weights);
-    static void writeCoordinates(std::string filePath, const std::vector<std::vector<double>>& positions);
+    static void writeCoordinates(std::string filePath, const std::vector<std::vector<float>>& positions,
+                                 const std::vector<float>& weights);
+    static void writeCoordinates(std::string filePath, const std::vector<std::vector<float>>& positions);
 };
