+++ src/embeddingLib/include/embeddingSpace/Additive.hpp
@@ -8,16 +8,16 @@
 */
 class Additive : public Embedding {
    public:
-    Additive(const std::vector<std::vector<double>> &coords, const std::vector<double> &weights);
+    Additive(const std::vector<std::vector<float>> &coords, const std::vector<float> &weights);
     virtual ~Additive(){};
 
-    virtual double getSimilarity(NodeId a, NodeId b) const;
+    virtual float getSimilarity(NodeId a, NodeId b) const;
     virtual int getDimension() const;
-    double getDistance(NodeId a, NodeId b) const;
-    double getNodeWeight(NodeId a) const;
+    float getDistance(NodeId a, NodeId b) const;
+    float getNodeWeight(NodeId a) const;
 
    private:
     const int DIMENSION;
     VecList coordinates;
-    std::vector<double> weights;
+    std::vector<float> weights;
 };
