+++ src/embeddingLib/include/embeddingSpace/Cosine.hpp
@@ -5,8 +5,8 @@
 
 class Cosine : public Embedding {
    public:
-    Cosine(const std::vector<std::vector<double>> &coords);
-    virtual double getSimilarity(NodeId a, NodeId b) const;
+    Cosine(const std::vector<std::vector<float>> &coords);
+    virtual float getSimilarity(NodeId a, NodeId b) const;
     virtual int getDimension() const;
 
    private:
