+++ src/embeddingLib/include/embeddingSpace/Embedding.hpp
@@ -15,5 +15,5 @@ class Embedding {
      * CAUTION: A low values indicates a high similarity. A high value indicates a low similarity.
      * Think of it like the euclidean distance.
     */
-    virtual double getSimilarity(NodeId a, NodeId b) const = 0;
+    virtual float getSimilarity(NodeId a, NodeId b) const = 0;
 };
