+++ src/embeddingLib/include/embeddingSpace/Euclidean.hpp
@@ -5,8 +5,8 @@
 
 class Euclidean : public Embedding {
    public:
-    Euclidean(const std::vector<std::vector<double>> &coords);
-    virtual double getSimilarity(NodeId a, NodeId b) const;
+    Euclidean(const std::vector<std::vector<float>> &coords);
+    virtual float getSimilarity(NodeId a, NodeId b) const;
     virtual int getDimension() const;
 
    private:
