+++ src/embeddingLib/include/embeddingSpace/WeightedNoDim.hpp
@@ -9,16 +9,16 @@
 */
 class WeightedNoDim : public Embedding {
    public:
-    WeightedNoDim(const std::vector<std::vector<double>> &coords, const std::vector<double> &weights);
+    WeightedNoDim(const std::vector<std::vector<float>> &coords, const std::vector<float> &weights);
     virtual ~WeightedNoDim(){};
 
-    virtual double getSimilarity(NodeId a, NodeId b) const;
+    virtual float getSimilarity(NodeId a, NodeId b) const;
     virtual int getDimension() const;
-    double getDistance(NodeId a, NodeId b) const;
-    double getNodeWeight(NodeId a) const;
+    float getDistance(NodeId a, NodeId b) const;
+    float getNodeWeight(NodeId a) const;
 
    private:
     const int DIMENSION;
     VecList coordinates;
-    std::vector<double> weights;
+    std::vector<float> weights;
 };
